# schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NPCs AI Studio · 本地数据服务",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/workspace/model-profiles": {
      "put": {
        "summary": "Provider Profile",
        "operationId": "provider_profile_api_workspace_model_profiles_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderProfile"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/models/test": {
      "post": {
        "summary": "Test Connection",
        "operationId": "test_connection_api_models_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionTestInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/projects/{project_id}/generate": {
      "post": {
        "summary": "Generate",
        "operationId": "generate_api_v2_projects__project_id__generate_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateInput"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/projects/{project_id}/generation": {
      "get": {
        "summary": "Generation Jobs",
        "operationId": "generation_jobs_api_v2_projects__project_id__generation_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/files/new": {
      "post": {
        "summary": "New File",
        "operationId": "new_file_api_files_new_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiskProjectInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CardTrial": {
        "properties": {
          "dialogue_id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,99}$",
            "title": "Dialogue Id"
          },
          "start_node_id": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,99}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Node Id"
          },
          "use_entry_routes": {
            "type": "boolean",
            "title": "Use Entry Routes",
            "default": false
          },
          "started": {
            "type": "boolean",
            "title": "Started",
            "default": false
          },
          "actions": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/TrialVariable"
                },
                {
                  "$ref": "#/components/schemas/TrialChoice"
                }
              ],
              "discriminator": {
                "propertyName": "type",
                "mapping": {
                  "choice": "#/components/schemas/TrialChoice",
                  "variable": "#/components/schemas/TrialVariable"
                }
              }
            },
            "type": "array",
            "maxItems": 128,
            "title": "Actions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "dialogue_id"
        ],
        "title": "CardTrial"
      },
...
      "ChoiceRecord": {
        "properties": {
          "tick": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Tick"
          },
          "dialogue_id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,99}$",
            "title": "Dialogue Id"
          },
          "option_id": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,99}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Option Id"
          },
          "action": {
            "type": "string",
            "enum": [
              "choice",
              "start",
              "restart"
            ],
            "title": "Action",
            "default": "choice"
          },
          "record_id": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,99}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Record Id"
          },
          "sequence": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Sequence"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "tick",
          "dialogue_id"
        ],
        "title": "ChoiceRecord"
      },
...
      "ConnectionTestInput": {
        "properties": {
          "profile": {
            "$ref": "#/components/schemas/ProviderProfile"
          },
          "api_key": {
            "type": "string",
            "format": "password",
            "title": "Api Key",
            "default": "",
            "writeOnly": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "profile"
        ],
        "title": "ConnectionTestInput"
      },
...
      "DiskProjectInput": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Name"
          },
          "folder": {
            "type": "string",
            "title": "Folder"
          },
          "filename": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Filename"
          },
          "template": {
            "type": "string",
            "enum": [
              "blank",
              "lighthouse",
              "outpost",
              "campfire"
            ],
            "title": "Template",
            "default": "blank"
          },
          "layout": {
            "type": "string",
            "enum": [
              "file",
              "folder"
            ],
            "title": "Layout",
            "default": "file"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "folder",
          "filename"
        ],
        "title": "DiskProjectInput"
      },
...
      "EntityRef": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "character",
              "location",
              "faction",
              "fact",
              "variable",
              "event",
              "rule",
              "dialogue",
              "text"
            ],
            "title": "Kind"
          },
          "id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,99}$",
            "title": "Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "id"
        ],
        "title": "EntityRef"
      },
...
      "GenerateInput": {
        "properties": {
          "request_id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,99}$",
            "title": "Request Id"
          },
          "expected_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Revision"
          },
          "profile_id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,99}$",
            "title": "Profile Id"
          },
          "purpose": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "character",
                  "story",
                  "dialogue",
                  "text"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Purpose"
          },
          "api_key": {
            "type": "string",
            "format": "password",
            "title": "Api Key",
            "default": "",
            "writeOnly": true
          },
          "instructions": {
            "type": "string",
            "maxLength": 30000,
            "title": "Instructions"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/GenerationItem"
            },
            "type": "array",
            "maxItems": 10,
            "minItems": 1,
            "title": "Items"
          },
          "source_refs": {
            "items": {
              "$ref": "#/components/schemas/EntityRef"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Source Refs"
          },
          "case_id": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,99}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Id"
          },
          "at_tick": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "At Tick"
          },
          "scenario": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SimulationInput"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "expected_revision",
          "profile_id",
          "instructions",
          "items"
        ],
        "title": "GenerateInput"
      },
      "GenerationItem": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "character",
              "dialogue",
              "text"
            ],
            "title": "Kind",
            "default": "character"
          },
          "name": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Name"
          },
          "target_id": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,99}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Id"
          },
          "character_id": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,99}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Character Id"
          },
          "fields": {
            "items": {
              "type": "string",
              "maxLength": 150,
              "minLength": 1
            },
            "type": "array",
            "maxItems": 30,
            "title": "Fields"
          },
          "scene_id": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,99}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scene Id"
          },
          "start_tick": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Tick"
          },
          "end_tick": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "End Tick"
          },
          "scene_context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SceneGenerationContext"
              },
              {
                "type": "null"
              }
            ]
          },
          "output_limits": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GenerationLimits"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "GenerationItem"
      },
      "GenerationLimits": {
        "properties": {
          "card_count": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 6.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Card Count"
          },
          "text_length": {
            "type": "string",
            "enum": [
              "short",
              "medium",
              "long"
            ],
            "title": "Text Length",
            "default": "short"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "GenerationLimits"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelTestResult": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "success",
              "failed"
            ],
            "title": "Status"
          },
          "checked_at": {
            "type": "string",
            "format": "date-time",
            "title": "Checked At"
          },
          "message": {
            "type": "string",
            "maxLength": 30000,
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "checked_at",
          "message"
        ],
        "title": "ModelTestResult"
      },
...
      "ProviderProfile": {
        "properties": {
          "endpoint": {
            "type": "string",
            "maxLength": 30000,
            "title": "Endpoint",
            "default": ""
          },
          "model": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 150,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,99}$",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Name"
          },
          "mode": {
            "type": "string",
            "enum": [
              "remote",
              "local"
            ],
            "title": "Mode",
            "default": "remote"
          },
          "protocol": {
            "type": "string",
            "const": "chat_completions",
            "title": "Protocol",
            "default": "chat_completions"
          },
          "stream": {
            "type": "boolean",
            "title": "Stream",
            "default": false
          },
          "json_mode": {
            "type": "boolean",
            "title": "Json Mode",
            "default": false
          },
          "timeout_seconds": {
            "type": "integer",
            "maximum": 300.0,
            "minimum": 5.0,
            "title": "Timeout Seconds",
            "default": 90
          },
          "max_tokens": {
            "type": "integer",
            "maximum": 16384.0,
            "minimum": 128.0,
            "title": "Max Tokens",
            "default": 2048
          },
          "retry_limit": {
            "type": "integer",
            "maximum": 2.0,
            "minimum": 0.0,
            "title": "Retry Limit",
            "default": 0
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "archived": {
            "type": "boolean",
            "title": "Archived",
            "default": false
          },
          "last_test": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelTestResult"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "ProviderProfile"
      },
...
      "SceneGenerationContext": {
        "properties": {
          "level_id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,99}$",
            "title": "Level Id"
          },
          "track_id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,99}$",
            "title": "Track Id"
          },
          "location_id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,99}$",
            "title": "Location Id"
          },
          "appearance_id": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,99}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Appearance Id"
          },
          "group_name": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Group Name"
          },
          "mode": {
            "type": "string",
            "enum": [
              "pool",
              "people"
            ],
            "title": "Mode"
          },
          "start_tick": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Start Tick"
          },
          "end_tick": {
            "type": "integer",
            "minimum": 0.0,
            "title": "End Tick"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "level_id",
          "track_id",
          "location_id",
          "group_name",
          "mode",
          "start_tick",
          "end_tick"
        ],
        "title": "SceneGenerationContext"
      },
      "SceneRecord": {
        "properties": {
          "sequence": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Sequence"
          },
          "tick": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Tick"
          },
          "location_id": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,99}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "tick"
        ],
        "title": "SceneRecord"
      },
...
      "SimulationInput": {
        "properties": {
          "expected_content_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Content Revision"
          },
          "case_id": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,99}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Id"
          },
          "at_tick": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "At Tick"
          },
          "location_id": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,99}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Id"
          },
          "level_id": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,99}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level Id"
          },
          "variable_overrides": {
            "patternProperties": {
              "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,99}$": {
                "anyOf": [
                  {
                    "type": "boolean"
                  },
                  {
                    "type": "integer"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "string"
                  }
                ]
              }
            },
            "type": "object",
            "title": "Variable Overrides"
          },
          "choices": {
            "items": {
              "$ref": "#/components/schemas/ChoiceRecord"
            },
            "type": "array",
            "maxItems": 10000,
            "title": "Choices"
          },
          "scene_changes": {
            "items": {
              "$ref": "#/components/schemas/SceneRecord"
            },
            "type": "array",
            "maxItems": 1000,
            "title": "Scene Changes"
          },
          "card_trial": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CardTrial"
              },
              {
                "type": "null"
              }
            ]
          },
          "max_steps": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 1.0,
            "title": "Max Steps",
            "default": 2048
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_content_revision"
        ],
        "title": "SimulationInput"
      },
...
      "TrialVariable": {
        "properties": {
          "type": {
            "type": "string",
            "const": "variable",
            "title": "Type"
          },
          "variable_id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,99}$",
            "title": "Variable Id"
          },
          "value": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "integer"
              },
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "title": "Value"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "type",
          "variable_id",
          "value"
        ],
        "title": "TrialVariable"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}